refactor model package
