+++ src/command/handler/simple-command-handler.ts
@@ -1,4 +1,4 @@
-import { CommandType } from "../../model/command-type";
+import { CommandType } from "../../model/api/command-type";
 import { CommandHandler } from "./command-handler";
 import shelljs from 'shelljs';
