+++ src/model/api/completion.ts
