+++ src/util/config-loader.ts
@@ -2,8 +2,8 @@ import fs from 'fs';
 import path from 'path';
 import walkdir from 'walkdir';
 import yaml from 'js-yaml';
-import { Alias } from '../model/alias';
-import { Config } from '../model/config';
+import { Alias } from '../model/api/alias';
+import { Config } from '../model/api/config';
 import { ConfigValidator } from '../validator/config-validator';
 import { AliasValidator } from '../validator/alias-validator';
 import '../extension/array-extensions';
