+++ src/validator/config-validator.ts
@@ -1,4 +1,4 @@
-import { Config } from "../model/config";
+import { Config } from "../model/api/config";
 
 export class ConfigValidator {
